Grouped routes into thematic traits. Added package structure.
