+++ eclair-node/src/main/scala/fr/acinq/eclair/api/directives/ErrorDirective.scala
@@ -1,8 +1,24 @@
+/*
+ * Copyright 2019 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package fr.acinq.eclair.api.directives
 
 import akka.http.scaladsl.model.{ContentTypes, HttpEntity, HttpResponse, StatusCodes}
 import akka.http.scaladsl.server.{Directive0, ExceptionHandler, RejectionHandler}
-import fr.acinq.eclair.api.{EclairDirectives, ErrorResponse, JsonSupport, Service}
+import fr.acinq.eclair.api.Service
 
 trait ErrorDirective {
   this: Service with EclairDirectives =>
@@ -15,7 +31,7 @@ trait ErrorDirective {
     handleRejections(apiRejectionHandler)
 
 
-  import JsonSupport.{formats, marshaller, serialization}
+  import fr.acinq.eclair.api.serde.JsonSupport.{formats, marshaller, serialization}
 
   private val apiExceptionHandler = ExceptionHandler {
     case t: IllegalArgumentException =>
