+++ eclair-node/src/main/scala/fr/acinq/eclair/api/directives/TimeoutDirective.scala
@@ -1,11 +1,27 @@
+/*
+ * Copyright 2019 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package fr.acinq.eclair.api.directives
 
 import akka.http.scaladsl.model.{ContentTypes, HttpRequest, HttpResponse, StatusCodes}
 import akka.http.scaladsl.server.{Directive0, Directive1, Directives}
-import fr.acinq.eclair.api.{ErrorResponse, JsonSupport}
 import akka.util.Timeout
-import fr.acinq.eclair.api.FormParamExtractors._
-import fr.acinq.eclair.api.JsonSupport._
+import fr.acinq.eclair.api.serde.FormParamExtractors._
+import fr.acinq.eclair.api.serde.JsonSupport._
+import fr.acinq.eclair.api.serde.JsonSupport
 
 import scala.concurrent.duration.DurationInt
