+++ eclair-node/src/main/scala/fr/acinq/eclair/api/handlers/WebSocket.scala
@@ -1,4 +1,20 @@
-package fr.acinq.eclair.api
+/*
+ * Copyright 2019 ACINQ SAS
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package fr.acinq.eclair.api.handlers
 
 import akka.NotUsed
 import akka.actor.{Actor, Props}
@@ -6,13 +22,14 @@ import akka.http.scaladsl.model.ws.{Message, TextMessage}
 import akka.http.scaladsl.server.Route
 import akka.stream.OverflowStrategy
 import akka.stream.scaladsl.{BroadcastHub, Flow, Keep, Source}
+import fr.acinq.eclair.api.Service
 import fr.acinq.eclair.channel.{ChannelClosed, ChannelCreated, ChannelStateChanged, WAIT_FOR_INIT_INTERNAL}
 import fr.acinq.eclair.payment.PaymentEvent
 
 trait WebSocket {
   this: Service =>
 
-  import JsonSupport.{formats, serialization}
+  import fr.acinq.eclair.api.serde.JsonSupport.{formats, serialization}
 
   /**
    * Http server route exposing the websocket handler
@@ -56,7 +73,6 @@ trait WebSocket {
       .merge(flowOutput) // Stream the data we want to the client
       .map(TextMessage.apply)
   }
-
 
 
 }
